+++ spb/__init__.py
@@ -49,7 +49,7 @@ from spb.graphics.functions_3d import (
     line_parametric_3d, list_3d, implicit_3d, wireframe, plane
 )
 from spb.graphics.vectors import (
-    vector_field_2d, vector_field_3d, arrow_2d
+    vector_field_2d, vector_field_3d, arrow_2d, arrow_3d
 )
 from spb.graphics.complex_analysis import (
     complex_points, line_abs_arg, line_abs_arg_colored, line_real_imag,
@@ -86,5 +86,5 @@ __all__ = [
     "hline", "vline",
     "pole_zero", "control_axis", "step_response", "impulse_response",
     "ramp_response", "bode_magnitude", "bode_phase", "nyquist", "nichols",
-    "arrow_2d"
+    "arrow_2d", "arrow_3d"
 ]