+++ spb/backends/matplotlib/renderers/__init__.py
@@ -14,11 +14,13 @@ from spb.backends.matplotlib.renderers.nichols import NicholsRenderer
 from spb.backends.matplotlib.renderers.arrow2d import (
     Arrow2DRendererQuivers, Arrow2DRendererFancyArrowPatch
 )
+from spb.backends.matplotlib.renderers.arrow3d import Arrow3DRendererFancyArrowPatch
 
 __all__ = [
     "Line2DRenderer", "Line3DRenderer", "ContourRenderer", "SurfaceRenderer",
     "GeometryRenderer", "Implicit2DRenderer", "Vector2DRenderer",
     "Vector3DRenderer", "ComplexRenderer", "GenericRenderer",
     "HVLineRenderer", "NyquistRenderer", "NicholsRenderer",
-    "Arrow2DRendererQuivers", "Arrow2DRendererFancyArrowPatch"
+    "Arrow2DRendererQuivers", "Arrow2DRendererFancyArrowPatch",
+    "Arrow3DRendererFancyArrowPatch"
 ]