+++ tests/graphics/test_vectors.py
@@ -9,7 +9,7 @@ from spb.series import (
 from sympy import (
     symbols, cos, sin, pi, sqrt, Matrix, tan, Tuple
 )
-from sympy.abc import j, k
+from sympy.abc import j, k, l
 from sympy.vector import CoordSys3D, gradient
 from sympy.physics.mechanics import ReferenceFrame
 
@@ -213,6 +213,38 @@ def test_arrow_2d(start, direc, label, rkw, sil, params):
     else:
         assert s.get_label(False) == (
             "(j, k) -> (j + 3, k + 4)" if not label else label)
+    assert s.rendering_kw == {} if not rkw else rkw
+    assert s.is_interactive == (len(s.params) > 0)
+    assert s.params == {} if not params else params
+
+
+@pytest.mark.parametrize(
+    "start, direc, label, rkw, sil, params",
+    [
+        ((1, 2, 3), (4, 5, 6), None, None, True, None),
+        ((1, 2, 3), (4, 5, 6), "test", {"color": "r"}, False, None),
+        (Tuple(j, k, l), (4, 5, 6), None, None, True, {j: (1, 0, 2), k: (2, 0, 3), l: (3, 0, 4)}),
+        (Tuple(j, k, l), (4, 5, 6), "test", {"color": "r"}, False, {j: (1, 0, 2), k: (2, 0, 3), l: (3, 0, 4)}),
+    ]
+)
+def test_arrow_3d(start, direc, label, rkw, sil, params):
+    kwargs = {}
+    if params:
+        kwargs["params"] = params
+    series = arrow_2d(
+        start, direc, label=label, rendering_kw=rkw, show_in_legend=sil,
+        **kwargs
+    )
+    assert len(series) == 1
+    s = series[0]
+    assert isinstance(s, Arrow2DSeries)
+    assert s.show_in_legend is sil
+    if not params:
+        assert s.get_label(False) == (
+            "(1.0, 2.0, 3.0) -> (5.0, 7.0, 9.0)" if not label else label)
+    else:
+        assert s.get_label(False) == (
+            "(j, k, l) -> (j + 4, k + 5, l + 6)" if not label else label)
     assert s.rendering_kw == {} if not rkw else rkw
     assert s.is_interactive == (len(s.params) > 0)
     assert s.params == {} if not params else params